Refactoring Tailwind and leveraging Webpacker
We didn't actually make any config changes to Tailwind or Webpacker
because we had already made the changes this section mentions, earlier
in the tutorial. The files referenced in this section are already in
sync. The only changes here are to add some new scss files to get ready
for upcoming work creating custom tailwind components.