UI Enhancements
updates to enhance look and function of drop down menus

Note: there is a discrepency with application.html.erb in this commit
and what is found in the tutorial. The tutorial shows a line for one of
<li> elements as: link_to profile_path(current_user). There is some
missing code here becuase that route has not yet been defined. Back in
module 4 layouts - the last place where we changed application.html.erb -
this looks like: link_to "#" which is what it is set to for now for this
commit. Otherwise it breaks the application with a NoMethodError. I've
let Andy know about this.