Checking our checks thru Authenticating Controllers with Devise
After the Update Community Views section the app breaks due to several
changes which are introduced and not resolved until the end of the module.
Instead of committing code for sections that would lead to the app being in
a broken state, I have opted to commit all the remaining changes for Module 8.